+++ umich-oidc-login/CHANGELOG.md
@@ -7,6 +7,7 @@ June 22, 2025
 * Changes since the private release of 1.3.0-beta3:
 	* Updated both README files, screenshots.
     * Added Settings and Support links for the plugin on the WordPress Plugins page.
+    * Fixed unstyled "Log in with Sigle Sign On" button on the WordPress login page.
     * Fixed "saving changes" spinner for access metabox.
     * Fixed location of GitHub updater files.
 	* Updated dependencies and tested up to WordPress 6.8.1.